# Use std::countl_zero instead of __builtin_clz